all pydantic classes moved to the interface
