+++ flood_adapt/object_model/direct_impact/measure/elevate.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 from typing import Any, Union
 
 import tomli
@@ -6,37 +7,37 @@ import tomli_w
 
 from flood_adapt.object_model.direct_impact.measure.impact_measure import (
     ImpactMeasure,
-    ImpactMeasureModel,
 )
-from flood_adapt.object_model.interface.measures import IMeasure
-from flood_adapt.object_model.io.unitfulvalue import UnitfulLengthRefValue
+from flood_adapt.object_model.interface.measures import ElevateModel, IElevate
 
 
-class ElevateModel(ImpactMeasureModel):
-    elevation: UnitfulLengthRefValue
-
-
-class Elevate(ImpactMeasure, IMeasure):
+class Elevate(ImpactMeasure, IElevate):
     """Subclass of ImpactMeasure describing the measure of elevating buildings by a specific height"""
 
     attrs: ElevateModel
+    database_input_path: Union[str, os.PathLike]
 
     @staticmethod
-    def load_file(filepath: Union[str, os.PathLike]):
+    def load_file(filepath: Union[str, os.PathLike]) -> IElevate:
         """create Elevate from toml file"""
 
         obj = Elevate()
         with open(filepath, mode="rb") as fp:
             toml = tomli.load(fp)
         obj.attrs = ElevateModel.parse_obj(toml)
+        # if measure is created by path use that to get to the database path
+        obj.database_input_path = Path(filepath).parents[2]
         return obj
 
     @staticmethod
-    def load_dict(data: dict[str, Any]):
+    def load_dict(
+        data: dict[str, Any], database_input_path: Union[str, os.PathLike]
+    ) -> IElevate:
         """create Elevate from object, e.g. when initialized from GUI"""
 
         obj = Elevate()
         obj.attrs = ElevateModel.parse_obj(data)
+        obj.database_input_path = database_input_path
         return obj
 
     def save(self, filepath: Union[str, os.PathLike]):
