+++ flood_adapt/object_model/direct_impact/socio_economic_change.py
@@ -1,17 +1,4 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
-from flood_adapt.object_model.io.unitfulvalue import UnitfulLengthRefValue
-
-
-class SocioEconomicChangeModel(BaseModel):
-    population_growth_existing: Optional[float] = 0.0
-    economic_growth: Optional[float] = 0.0
-
-    population_growth_new: Optional[float] = 0.0
-    new_development_elevation: Optional[UnitfulLengthRefValue]
-    new_development_shapefile: Optional[str]
+from flood_adapt.object_model.interface.projections import SocioEconomicChangeModel
 
 
 class SocioEconomicChange:
