+++ flood_adapt/object_model/hazard/hazard_strategy.py
@@ -2,7 +2,12 @@ from flood_adapt.object_model.hazard.measure.hazard_measure import HazardMeasure
 
 
 class HazardStrategy:
-    """Subclass of Strategy describing a strategy with only hazard measures"""
+    """Class containing only the hazard measures of a strategy"""
 
     def __init__(self, measures: list[HazardMeasure]) -> None:
+        """
+        Parameters
+        ----------
+        measures : list[HazardMeasure]
+        """
         self.measures = measures
