+++ flood_adapt/object_model/interface/projections.py
@@ -1,9 +1,41 @@
 import os
 from abc import ABC, abstractmethod
-from typing import Any, Union
+from typing import Any, Optional, Union
+
+from pydantic import BaseModel
+
+from flood_adapt.object_model.io.unitfulvalue import (
+    UnitfulLength,
+    UnitfulLengthRefValue,
+)
+
+
+class PhysicalProjectionModel(BaseModel):
+    sea_level_rise: Optional[UnitfulLength] = UnitfulLength(value=0.0, units="meters")
+    subsidence: Optional[UnitfulLength] = UnitfulLength(value=0.0, units="meters")
+    rainfall_increase: Optional[float] = 0.0
+    storm_frequency_increase: Optional[float] = 0.0
+
+
+class SocioEconomicChangeModel(BaseModel):
+    population_growth_existing: Optional[float] = 0.0
+    economic_growth: Optional[float] = 0.0
+
+    population_growth_new: Optional[float] = 0.0
+    new_development_elevation: Optional[UnitfulLengthRefValue]
+    new_development_shapefile: Optional[str]
+
+
+class ProjectionModel(BaseModel):
+    name: str
+    long_name: str
+    physical_projection: PhysicalProjectionModel
+    socio_economic_change: SocioEconomicChangeModel
 
 
 class IProjection(ABC):
+    attrs: ProjectionModel
+
     @staticmethod
     @abstractmethod
     def load_file(filepath: Union[str, os.PathLike]):
