+++ flood_adapt/object_model/interface/scenarios.py
@@ -2,8 +2,22 @@ import os
 from abc import ABC, abstractmethod
 from typing import Any, Union
 
+from pydantic import BaseModel
+
+
+class ScenarioModel(BaseModel):
+    """BaseModel describing the expected variables and data types of a scenario"""
+
+    name: str
+    long_name: str
+    event: str
+    projection: str
+    strategy: str
+
 
 class IScenario(ABC):
+    attrs: ScenarioModel
+
     @staticmethod
     @abstractmethod
     def load_file(filepath: Union[str, os.PathLike]):
@@ -12,7 +26,7 @@ class IScenario(ABC):
 
     @staticmethod
     @abstractmethod
-    def load_dict(data: dict[str, Any]):
+    def load_dict(data: dict[str, Any], database_input_path: Union[str, os.PathLike]):
         """get Scenario attributes from an object, e.g. when initialized from GUI"""
         ...
