+++ flood_adapt/object_model/measure.py
@@ -1,22 +1,16 @@
-from pathlib import Path
+import os
+from typing import Union
 
 import tomli
-from pydantic import BaseModel
 
-
-class MeasureModel(BaseModel):
-    """"""
-
-    name: str
-    long_name: str
-    type: str
+from flood_adapt.object_model.interface.measures import MeasureModel
 
 
 class Measure:
     attrs: MeasureModel
 
     @staticmethod
-    def load_measure_type(filepath: Path):
+    def get_measure_type(filepath: Union[str, os.PathLike]):
         """get a measure type from toml file"""
 
         obj = Measure()
